+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.2 (WIP)
 **New features:**
+- Added support for resizing the terminal
 - Added keybindings: <kbd>g</kbd> / <kbd>G</kbd> to go to first and last element
 - Added keybinding: <kbd>c</kbd> to open submission comments in browser